+++ src/custom/RLSearchLayer.cpp
@@ -118,7 +118,7 @@ bool RLSearchLayer::init() {
       m_searchInputMenu->setPosition({winSize.width / 2, winSize.height - 30});
       m_searchInputMenu->setContentSize({365.f, 40.f});
 
-      auto searchInputBg = CCScale9Sprite::create("square02_001.png");
+      auto searchInputBg = CCScale9Sprite::create("square02_small.png");
       searchInputBg->setContentSize(m_searchInputMenu->getContentSize());
       searchInputBg->setOpacity(100);
       m_searchInputMenu->addChild(searchInputBg, -1);
@@ -216,7 +216,7 @@ bool RLSearchLayer::init() {
             m_difficultyMenuItems.push_back(menuItem);
             m_difficultySprites.push_back(difficultySprite);
       }
-      auto difficultyMenuBg = CCScale9Sprite::create("square02_001.png");
+      auto difficultyMenuBg = CCScale9Sprite::create("square02_small.png");
       difficultyMenuBg->setContentSize(m_difficultyFilterMenu->getContentSize());
       difficultyMenuBg->setPosition(m_difficultyFilterMenu->getPosition());
       difficultyMenuBg->setOpacity(100);
@@ -313,7 +313,7 @@ bool RLSearchLayer::init() {
       optionsLabel->setPosition({winSize.width / 2, winSize.height - 160});
       optionsLabel->setScale(0.5f);
       this->addChild(optionsLabel);
-      auto optionsMenuBg = CCScale9Sprite::create("square02_001.png");
+      auto optionsMenuBg = CCScale9Sprite::create("square02_small.png");
       optionsMenuBg->setContentSize(optionsMenu->getContentSize());
       optionsMenuBg->setPosition(optionsMenu->getPosition());
       optionsMenuBg->setOpacity(100);